refactor : test용 클래스 위치 옮김
