+++ src/main/java/IDMaker/IDMakerProject.java
@@ -1,4 +1,4 @@
-package IDMaker.idmaker;
+package IDMaker;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
