+++ src/main/java/IDMaker/fixture/repository/TestClassRepository.java
@@ -1,8 +1,8 @@
-package IDMaker.idmaker.fixture.repository;
+package IDMaker.fixture.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import IDMaker.idmaker.fixture.entity.TestClass;
+import IDMaker.fixture.entity.TestClass;
 
 public interface TestClassRepository extends JpaRepository<TestClass, String> {
 }
